ai_engine/requirements.txt
flask
pandas
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv